+++ xarray/tests/test_accessors.py
@@ -161,6 +161,8 @@ def times_3d(times):
 @pytest.mark.parametrize('field', ['year', 'month', 'day', 'hour',
                                    'dayofyear', 'dayofweek'])
 def test_field_access(data, field):
+    if field == 'dayofyear' or field == 'dayofweek':
+        pytest.importorskip('cftime', minversion='1.0.2.1')
     result = getattr(data.time.dt, field)
     expected = xr.DataArray(
         getattr(xr.coding.cftimeindex.CFTimeIndex(data.time.values), field),
@@ -176,6 +178,8 @@ def test_field_access(data, field):
 def test_dask_field_access_1d(data, field):
     import dask.array as da
 
+    if field == 'dayofyear' or field == 'dayofweek':
+        pytest.importorskip('cftime', minversion='1.0.2.1')
     expected = xr.DataArray(
         getattr(xr.coding.cftimeindex.CFTimeIndex(data.time.values), field),
         name=field, dims=['time'])
@@ -193,6 +197,8 @@ def test_dask_field_access_1d(data, field):
 def test_dask_field_access(times_3d, data, field):
     import dask.array as da
 
+    if field == 'dayofyear' or field == 'dayofweek':
+        pytest.importorskip('cftime', minversion='1.0.2.1')
     expected = xr.DataArray(
         getattr(xr.coding.cftimeindex.CFTimeIndex(times_3d.values.ravel()),
                 field).reshape(times_3d.shape),